Switch to releases based on Heroku stack
Previously all builds were against Cedar, which is probably fine, but
building against newer stacks may give benefits from a new toolchain.
This switches to a strategy of having a release we stack that contains
all of the specific version of jemalloc.

This also drops the additional build version number (e.g. 5.0.1-1) as
it's not configurable by the user and could be different per version.
Anyone using the build pack will always get the newest build of a
particular version that's available.