Switch to using symlink to jemalloc in LD_PRELOAD
An issue came up that jemalloc 3.6.0 does not include jemalloc-config
(it was added in 4.0.0). Researching the issue I discovered that
installing jemalloc will create a symlink that points to the correct
revision. Using jemalloc-config is more correct especially in cases
where multiple versions of jemalloc are installed. Since Heroku dyno's
are built from a fresh state on each code change, we can expect that
there will only ever be one version of jemalloc installed and the
symlink will point to the latest version.